move TimerHttpMiddleware to seperate file
