+++ src/FluentlyHttpClient/LoggerHttpMiddleware.cs
@@ -1,10 +1,12 @@
 ﻿using Microsoft.Extensions.Logging;
 using System;
-using System.Diagnostics;
 using System.Threading.Tasks;
 
 namespace FluentlyHttpClient
 {
+	/// <summary>
+	/// Logging middleware for Http client.
+	/// </summary>
 	public class LoggerHttpMiddleware : IFluentHttpMiddleware
 	{
 		private readonly FluentHttpRequestDelegate _next;
@@ -25,34 +27,6 @@ namespace FluentlyHttpClient
 
 			if (_logger.IsEnabled(LogLevel.Information))
 				_logger.LogInformation("Post-request... {response}", response);
-			return response;
-		}
-	}
-	public class TimerHttpMiddleware : IFluentHttpMiddleware
-	{
-		private readonly FluentHttpRequestDelegate _next;
-		private readonly ILogger _logger;
-
-		public TimerHttpMiddleware(FluentHttpRequestDelegate next, ILogger<TimerHttpMiddleware> logger)
-		{
-			_next = next;
-			_logger = logger;
-		}
-
-		public async Task<FluentHttpResponse> Invoke(FluentHttpRequest request)
-		{
-			var watch = Stopwatch.StartNew();
-			var response = await _next(request);
-			var elapsed = watch.Elapsed;
-
-			// todo: make configurable
-			const int thresholdMillis = 250;
-			if (_logger.IsEnabled(LogLevel.Warning) && elapsed.TotalMilliseconds >= thresholdMillis)
-				_logger.LogWarning("Executed request {request} in {timeTakenMillis}ms", request, elapsed.TotalMilliseconds);
-			else if (_logger.IsEnabled(LogLevel.Information))
-				_logger.LogInformation("Executed request {request} in {timeTakenMillis}ms", request, elapsed.TotalMilliseconds);
-
-			response.SetTimeTaken(elapsed);
 			return response;
 		}
 	}
